Changes made at tournament to make bot better
